+++ client/src/index.js
@@ -14,11 +14,9 @@
  limitations under the License.
  */
 
-import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./app/App";
-import registerServiceWorker from "./registerServiceWorker";
 import { BrowserRouter as Router } from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
@@ -28,5 +26,3 @@ root.render(
         <App />
     </Router>
 );
-
-// registerServiceWorker();